pyproject.toml
[project]
name = "cinereserve"
version = "0.1.0"
description = ""
authors = [
    {name = "João Lucas de Melo",email = "joaolucasmm2512@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "redis (>=7.3.0,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "django-stubs[compatible-mypy] (>=6.0.0,<7.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "djangorestframework-stubs[compatible-mypy] (>=3.16.8,<4.0.0)"
]

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[tool.django-stubs]
django_settings_module = "cinereserve.settings"